move TestRegionInfoBuilder to hbase-client, leave a TestReadAndWriteRegionInfoFile in hbase-server
